Flatten types and clean up module exports (#103)
This commit introduces multiple changes:

- It removes the wrapper types we historically had around Config and Client.
  AFAICT, these are redundant and pointless.

- It moves from context functions into an internal module. This allows me to
  remove 5 functions from the main Context module export since we don't
  actually want this to be a part of the public API.

- Updates the main Server module to take advantage of full module re-exports
  since that is all we are doing. This prevents us from having to update export
  lists in two places.

- Removes the now defunct user module.
