+++ src/LaunchDarkly/Server/Client/Internal.hs
@@ -1,6 +1,5 @@
 module LaunchDarkly.Server.Client.Internal
     ( Client (..)
-    , ClientI (..)
     , Status (..)
     , clientVersion
     , setStatus
@@ -15,26 +14,34 @@ import Data.Text (Text)
 import GHC.Generics (Generic)
 
 import LaunchDarkly.Server.Client.Status (Status (..), transitionStatus)
-import LaunchDarkly.Server.Config.Internal (ConfigI)
+import LaunchDarkly.Server.Config.Internal (Config)
 import LaunchDarkly.Server.DataSource.Internal (DataSource)
 import LaunchDarkly.Server.Events (EventState)
 import LaunchDarkly.Server.Store.Internal (StoreHandle, getInitializedC)
-
--- |
--- Client is the LaunchDarkly client. Client instances are thread-safe.
--- Applications should instantiate a single instance for the lifetime of their
--- application.
-newtype Client = Client ClientI
 
 -- | The version string for this library.
 clientVersion :: Text
 clientVersion = "3.1.0"
 
-setStatus :: ClientI -> Status -> IO ()
+-- |
+-- Client is the LaunchDarkly client. Client instances are thread-safe.
+-- Applications should instantiate a single instance for the lifetime of their
+-- application.
+data Client = Client
+    { config :: !(Config)
+    , store :: !(StoreHandle IO)
+    , status :: !(IORef Status)
+    , events :: !EventState
+    , eventThreadPair :: !(Maybe (ThreadId, MVar ()))
+    , dataSource :: !DataSource
+    }
+    deriving (Generic)
+
+setStatus :: Client -> Status -> IO ()
 setStatus client status' =
     atomicModifyIORef' (getField @"status" client) (fmap (,()) (transitionStatus status'))
 
-getStatusI :: ClientI -> IO Status
+getStatusI :: Client -> IO Status
 getStatusI client =
     readIORef (getField @"status" client) >>= \case
         Unauthorized -> pure Unauthorized
@@ -43,13 +50,3 @@ getStatusI client =
             getInitializedC (getField @"store" client) >>= \case
                 Right True -> pure Initialized
                 _ -> pure Uninitialized
-
-data ClientI = ClientI
-    { config :: !(ConfigI)
-    , store :: !(StoreHandle IO)
-    , status :: !(IORef Status)
-    , events :: !EventState
-    , eventThreadPair :: !(Maybe (ThreadId, MVar ()))
-    , dataSource :: !DataSource
-    }
-    deriving (Generic)
