+++ test/Spec/Client.hs
@@ -18,8 +18,6 @@ import LaunchDarkly.Server.Features
 import LaunchDarkly.Server.Operators
 import LaunchDarkly.Server.Store
 import LaunchDarkly.Server.Store.Internal
-import LaunchDarkly.Server.User
-import LaunchDarkly.Server.User.Internal
 
 import Data.Text (Text)
 import LaunchDarkly.AesonCompat (fromList)
@@ -34,7 +32,7 @@ makeEmptyStore = do
 
 testSecureModeHashIsGeneratedCorrectly :: Test
 testSecureModeHashIsGeneratedCorrectly = TestCase $ do
-    client@(Client clientI) <- makeTestClient "secret"
+    client <- makeTestClient "secret"
     assertEqual "" "aa747c502a898200f9e4fa21bac68136f886a0e27aec70ba06daf2e2a5cb5597" (secureModeHash client userContext)
     assertEqual "" "a045e65c6d23bda4559ed4f2371a2508ce63016ceff58b00aa07b435e2bfedaa" (secureModeHash client orgContext)
   where
@@ -43,9 +41,9 @@ testSecureModeHashIsGeneratedCorrectly = TestCase $ do
 
 makeTestClient :: Text -> IO Client
 makeTestClient sdkKey = do
-    (Client client) <- makeClient $ (makeConfig sdkKey) & configSetOffline True
+    client <- makeClient $ (makeConfig sdkKey) & configSetOffline True
     initializeStore (getField @"store" client) mempty mempty
-    pure (Client client)
+    pure client
 
 allTests :: Test
 allTests =
